+++ src/handlePackages.js
@@ -6,14 +6,27 @@ function initPackageJson(config) {
 }
 
 function preparePackages(config, pkg) {
-  const isInstalled = isInstalledFactory(pkg);
-  const basicPackages = ['eslint', 'prettier', 'lint-staged', 'husky'];
-  let packagesToInstall = pkg
-    ? basicPackages.filter(item => config[item] && !isInstalled(item))
-    : basicPackages.filter(item => config[item]);
+  if (!pkg) {
+    throw new Error(
+      "package.json does not exist. Perhaps you don't have write access in this directory?"
+    );
+  }
 
-  if (packagesToInstall.includes('eslint') && packagesToInstall.includes('prettier')) {
-    packagesToInstall.push('eslint-config-prettier', 'eslint-plugin-prettier');
+  const isInstalled = isInstalledFactory(pkg);
+  const basicPackages = ['eslint', 'prettier', 'lint-staged', 'husky', 'stylelint'];
+  let packagesToInstall = basicPackages.filter(item => config[item] && !isInstalled(item));
+
+  if (packagesToInstall.includes('stylelint')) {
+    packagesToInstall.push('stylelint-config-recommended');
+  }
+
+  if (packagesToInstall.includes('prettier') || isInstalled('prettier')) {
+    if (packagesToInstall.includes('eslint')) {
+      packagesToInstall.push('eslint-config-prettier', 'eslint-plugin-prettier');
+    }
+    if (packagesToInstall.includes('stylelint')) {
+      packagesToInstall.push('stylelint-config-prettier', 'stylelint-prettier');
+    }
   }
 
   // CRA automatically installs eslint and most of the time doesn't work with manually installed one
@@ -78,10 +91,24 @@ function generateNewPackageJson(packagesToInstall, pkg) {
     if (!Array.isArray(arr)) arr = [arr];
     return arr[all ? 'every' : 'some'](i => packagesToInstall.includes(i));
   };
-  const isEslintInstalled = isInstalled('eslint') || usesCRA(pkg);
 
-  const prettierExts = 'htm,html,css,scss,less,graphql,json,md,yaml,yml';
   const eslintExts = 'ts,tsx,js,jsx';
+  const stylelintExts = 'css,scss,less';
+
+  const getPrettierExts = (isEslint, isStylelint) => {
+    let extensions = 'htm,html,graphql,json,md,yaml,yml';
+
+    if (isEslint) {
+      extensions = `${extensions},${eslintExts}`;
+    }
+    if (isStylelint) {
+      extensions += `${extensions},${stylelintExts}`;
+    }
+
+    return `*.{${extensions}}`;
+  };
+
+  const isEslintInstalled = isInstalled('eslint') || usesCRA(pkg);
 
   const config = {
     ...spreadConfig(
@@ -89,6 +116,12 @@ function generateNewPackageJson(packagesToInstall, pkg) {
       'eslintConfig',
       generateEslintConfig(packagesToInstall, pkg)
     ),
+    ...spreadConfig(!pkg.stylelint && isInstalled(['stylelint']), 'stylelint', {
+      extends: [
+        'stylelint-config-recommended',
+        ...(isInstalled(['prettier']) ? ['stylelint-prettier/recommended'] : [])
+      ]
+    }),
     ...spreadConfig(!pkg.husky && isInstalled(['husky', 'lint-staged'], true), 'husky', {
       hooks: {
         'pre-commit': 'lint-staged'
@@ -99,14 +132,18 @@ function generateNewPackageJson(packagesToInstall, pkg) {
       'lint-staged',
       {
         ...spreadConfig(isEslintInstalled, `*.{${eslintExts}}`, ['eslint --fix', 'git add']),
+        ...spreadConfig(isInstalled('stylelint'), `*.{${stylelintExts}}`, [
+          'stylelint --fix',
+          'git add'
+        ]),
         ...spreadConfig(
           isInstalled('prettier'),
-          isEslintInstalled ? `*.{${prettierExts}}` : `*.{${eslintExts},${prettierExts}}`,
+          getPrettierExts(isEslintInstalled, isInstalled('stylelint')),
           ['prettier --write', 'git add']
         )
       }
     ),
-    ...spreadConfig(!pkg.prettier && isInstalled(['prettier']), 'prettier', {
+    ...spreadConfig(!pkg.prettier && isInstalled('prettier'), 'prettier', {
       singleQuote: true
     })
   };