+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### Changelog
 
+#### [v0.3.0](https://github.com/paolostyle/autoslap/compare/v0.2.1...v0.3.0) (2020-03-28)
+
+- Added Stylelint support through a flag `--stylelint`, including configs for Prettier
+- Prettier configs and plugins for ESLint and Stylelint are now also installed and added to config if Prettier was already installed, but ESLint/Stylelint wasn't
+
 #### [v0.2.1](https://github.com/paolostyle/autoslap/compare/v0.2.0...v0.2.1) (2019-05-18)
 
 - Dropped jsonfile dependency, updated dependencies